add IncrementalCache
