+++ src/nixpacks/builder/docker/docker_image_builder.rs
@@ -1,19 +1,19 @@
 use super::{dockerfile_generation::DockerfileGenerator, DockerBuilderOptions, ImageBuilder};
 use crate::nixpacks::{
-    builder::docker::{dockerfile_generation::OutputDir, file_server::FileServer},
+    builder::docker::{
+        dockerfile_generation::OutputDir, file_server::FileServer,
+        incremental_cache::IncrementalCache,
+    },
     environment::Environment,
     files,
     logger::Logger,
     plan::BuildPlan,
 };
 use anyhow::{bail, Context, Ok, Result};
-use serde::{Deserialize, Serialize};
 use std::{
     fs::{self, remove_dir_all, File},
-    path::{Path, PathBuf},
     process::Command,
 };
-use tar::Archive;
 use tempdir::TempDir;
 use uuid::Uuid;
 
@@ -30,181 +30,6 @@ fn get_output_dir(app_src: &str, options: &DockerBuilderOptions) -> Result<Outpu
     } else {
         let tmp = TempDir::new("nixpacks").context("Creating a temp directory")?;
         OutputDir::new(tmp.into_path(), true)
-    }
-}
-
-fn write_incremental_cache_dockerfile(dir_path: &PathBuf) -> Result<PathBuf> {
-    let dockerfile_path = dir_path.join("Dockerfile");
-    if fs::metadata(&dockerfile_path).is_ok() {
-        fs::remove_file(&dockerfile_path).context("Remove old incremental cache dockerfile")?;
-    }
-
-    let paths = fs::read_dir(&dir_path)
-        .context("Read files at incremental cache dir")?
-        .filter_map(|path| {
-            path.ok()?
-                .file_name()
-                .to_str()
-                .map(|p| format!("COPY {} {}", p, p))
-        })
-        .collect::<Vec<_>>()
-        .join("\n");
-
-    let dockerfile = format!("FROM alpine\n{}", paths);
-
-    fs::write(dockerfile_path.clone(), dockerfile).context("Write incremental cache dockerfile")?;
-
-    Ok(dockerfile_path)
-}
-
-fn build_incremental_cache_image(dir_path: &PathBuf, tag: String) -> Result<()> {
-    let dockerfile_path = write_incremental_cache_dockerfile(dir_path)?;
-    let mut docker_build_cmd = Command::new("docker");
-
-    // Enable BuildKit for all builds
-    docker_build_cmd.env("DOCKER_BUILDKIT", "1");
-
-    docker_build_cmd
-        .arg("build")
-        .arg(&dir_path.display().to_string())
-        .arg("-f")
-        .arg(dockerfile_path.display().to_string())
-        .arg("-t")
-        .arg(tag);
-
-    let result = docker_build_cmd
-        .spawn()?
-        .wait()
-        .context("Build incremental cache image")?;
-
-    if !result.success() {
-        bail!("Building incremental cache image failed")
-    }
-
-    Ok(())
-}
-
-fn pull_incremental_cache_from_image(tag: &str) -> Result<()> {
-    let mut docker_build_cmd = Command::new("docker");
-
-    docker_build_cmd.arg("pull").arg(&tag);
-
-    let result = docker_build_cmd
-        .spawn()?
-        .wait()
-        .context("Pull incremental cache image")?;
-
-    if !result.success() {
-        bail!("Pulling incremental cache image failed")
-    }
-
-    Ok(())
-}
-
-fn save_incremental_cache_image_to_tar(tag: &str, file_path: &Path) -> Result<()> {
-    let mut docker_save_cmd = Command::new("docker");
-    docker_save_cmd
-        .arg("save")
-        .arg("-o")
-        .arg(file_path.display().to_string())
-        .arg(&tag);
-
-    let result = docker_save_cmd
-        .spawn()?
-        .wait()
-        .context("Pull incremental cache image")?;
-
-    if !result.success() {
-        bail!("Pulling incremental cache image failed")
-    }
-
-    Ok(())
-}
-
-fn download_incremental_cache_files(tag: &str, out_dir: &OutputDir) -> Result<()> {
-    let dir_path = out_dir.get_absolute_path("incremental-cache-image");
-    if fs::metadata(&dir_path)
-        .context("Check if incremental cache image dir exists")
-        .is_err()
-    {
-        fs::create_dir_all(&dir_path).context("Create incremental cache image dir")?;
-    }
-
-    let tar_file_path = dir_path.join("oci-image.tar");
-
-    pull_incremental_cache_from_image(tag)?;
-    save_incremental_cache_image_to_tar(tag, &tar_file_path)?;
-    let archives = extract_incremental_cache_image(
-        &tar_file_path,
-        &out_dir.get_absolute_path("incremental-cache-image"),
-    )?;
-
-    for item in archives {
-        let to_path = out_dir
-            .get_absolute_path("incremental-cache")
-            .join(item.name);
-        fs::rename(item.path, to_path).context("Move tar file to incremental cahce")?;
-    }
-
-    Ok(())
-}
-
-#[derive(Serialize, Deserialize)]
-struct Manifest {
-    #[serde(rename = "Layers")]
-    layers: Vec<String>,
-}
-struct IncrementalCacheArchive {
-    name: String,
-    path: PathBuf,
-}
-
-fn extract_incremental_cache_image(
-    file_path: &PathBuf,
-    dest_dir: &PathBuf,
-) -> Result<Vec<IncrementalCacheArchive>> {
-    let file = File::open(file_path)?;
-    let mut archive = Archive::new(file);
-    archive.unpack(&dest_dir)?;
-
-    let json_path = dest_dir.join("manifest.json");
-    let json_str = fs::read_to_string(json_path).context("Read manifest.json")?;
-    let value: Vec<Manifest> = serde_json::from_str(&json_str)?;
-
-    if value.first().is_none() {
-        Ok(vec![])
-    } else {
-        let mut archives: Vec<IncrementalCacheArchive> = vec![];
-        for layer_name in value.first().unwrap().layers.iter().skip(1) {
-            let tar_file_path = dest_dir.join(layer_name);
-            println!("layer_name {}", layer_name);
-
-            let extract_to = dest_dir.join(layer_name.replace("/layer.tar", "/layer"));
-            println!("extract_to {}", extract_to.display());
-
-            fs::create_dir_all(&extract_to).context("Create extract-to dir")?;
-            let file = File::open(tar_file_path)?;
-
-            let mut archive = Archive::new(file);
-            archive.unpack(&extract_to)?;
-
-            let mut found_files = fs::read_dir(&extract_to)
-                .context("Read files of extract-to dir")?
-                .filter_map(|path| {
-                    path.ok()?
-                        .file_name()
-                        .to_str()
-                        .map(std::string::ToString::to_string)
-                })
-                .map(|value| IncrementalCacheArchive {
-                    name: value.clone(),
-                    path: extract_to.join(value),
-                })
-                .collect::<Vec<_>>();
-
-            archives.append(&mut found_files);
-        }
-        Ok(archives)
     }
 }
 
@@ -228,23 +53,23 @@ impl ImageBuilder for DockerImageBuilder {
             return Ok(());
         }
 
-        if self.options.incremental_cache_image.is_some() {
-            let save_to = output.root.join(".nixpacks").join("incremental-cache");
-            if fs::metadata(&save_to)
-                .context("Check if incremental-cache dir exists")
-                .is_err()
-            {
-                fs::create_dir_all(&save_to).context("Creating incremental-cache directory")?;
-            }
+        let incremental_cache = IncrementalCache::default();
 
-            download_incremental_cache_files(
+        let incremental_cache_dirs = if self.options.incremental_cache_image.is_some() {
+            let dirs = incremental_cache.ensure_dirs_exists(&output)?;
+            incremental_cache.download_files(
                 &self.options.incremental_cache_image.clone().unwrap(),
-                &output,
+                &dirs,
             )?;
 
-            let file_receiver = FileServer::new(save_to, file_server_access_token);
-            file_receiver.start();
-        }
+            let file_server =
+                FileServer::new(dirs.tar_archives_dir.clone(), file_server_access_token);
+            file_server.start();
+
+            Some(dirs)
+        } else {
+            None
+        };
 
         println!("{}", plan.get_build_string()?);
 
@@ -273,10 +98,10 @@ impl ImageBuilder for DockerImageBuilder {
             println!("\nRun:");
             println!("  docker run -it {}", name);
 
-            if self.options.incremental_cache_image.is_some() {
+            if self.options.incremental_cache_image.is_some() && incremental_cache_dirs.is_some() {
                 println!("Building  incremental cache image!");
-                build_incremental_cache_image(
-                    &output.get_absolute_path("incremental-cache"),
+                incremental_cache.build_image(
+                    &incremental_cache_dirs.unwrap(),
                     self.options.incremental_cache_image.clone().unwrap(),
                 )?;
             }
